wsi-viewer-local/requirements.txt
flask~=3.1.0
flask-cors~=3.0.10
requests~=2.28.1
google-auth~=2.11.0
transformers>=4.40.0
torch>=2.1.0
pillow>=10.0.0
numpy>=1.24.0
ez-wsi-dicomweb~=6.0.9
